requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
SQLAlchemy>=2.0.20
pandas>=2.2.0
qrcode>=8.0
Pillow>=10.0.0
reportlab>=4.0.0
python-dotenv>=1.0.0
Werkzeug>=3.0.0
